Check for known services before external IPs
Known services can be internal (e.g. same VPC in AWS)
